requirements.txt
telethon
python-dotenv
psycopg2-binary
fastapi
uvicorn
sqlalchemy
dagster
dagster-webserver
